+++ packages/std/src/exports.rs
@@ -5,7 +5,6 @@
 //!
 //! do_init and do_wrapper should be wrapped with a extern "C" entry point
 //! including the contract-specific init/handle function pointer.
-use std::fmt::Display;
 use std::os::raw::c_void;
 use std::vec::Vec;
 
@@ -53,10 +52,9 @@ pub fn do_init<T: DeserializeOwned + JsonSchema>(
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
 ) -> *mut c_void {
-    match _do_init(init_fn, env_ptr, msg_ptr) {
-        Ok(res) => res,
-        Err(err) => make_init_error_c_string(err),
-    }
+    let res: InitResult = _do_init(init_fn, env_ptr, msg_ptr).into();
+    let v = to_vec(&res).unwrap();
+    release_buffer(v)
 }
 
 /// do_handle should be wrapped in an external "C" export, containing a contract-specific function as arg
@@ -69,10 +67,9 @@ pub fn do_handle<T: DeserializeOwned + JsonSchema>(
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
 ) -> *mut c_void {
-    match _do_handle(handle_fn, env_ptr, msg_ptr) {
-        Ok(res) => res,
-        Err(err) => make_handle_error_c_string(err),
-    }
+    let res: HandleResult = _do_handle(handle_fn, env_ptr, msg_ptr).into();
+    let v = to_vec(&res).unwrap();
+    release_buffer(v)
 }
 
 /// do_query should be wrapped in an external "C" export, containing a contract-specific function as arg
@@ -80,10 +77,9 @@ pub fn do_query<T: DeserializeOwned + JsonSchema>(
     query_fn: &dyn Fn(&Extern<ExternalStorage, ExternalApi>, T) -> Result<QueryResponse, Error>,
     msg_ptr: *mut c_void,
 ) -> *mut c_void {
-    match _do_query(query_fn, msg_ptr) {
-        Ok(res) => res,
-        Err(err) => make_query_error_c_string(err),
-    }
+    let res: QueryResult = _do_query(query_fn, msg_ptr).into();
+    let v = to_vec(&res).unwrap();
+    release_buffer(v)
 }
 
 fn _do_init<T: DeserializeOwned + JsonSchema>(
@@ -94,15 +90,13 @@ fn _do_init<T: DeserializeOwned + JsonSchema>(
     ) -> Result<InitResponse, Error>,
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
-) -> Result<*mut c_void, Error> {
+) -> Result<InitResponse, Error> {
     let env: Vec<u8> = unsafe { consume_region(env_ptr)? };
     let msg: Vec<u8> = unsafe { consume_region(msg_ptr)? };
     let env: Env = from_slice(&env)?;
     let msg: T = from_slice(&msg)?;
     let mut deps = make_dependencies();
-    let res = init_fn(&mut deps, env, msg)?;
-    let json = to_vec(&InitResult::Ok(res))?;
-    Ok(release_buffer(json))
+    init_fn(&mut deps, env, msg)
 }
 
 fn _do_handle<T: DeserializeOwned + JsonSchema>(
@@ -113,44 +107,25 @@ fn _do_handle<T: DeserializeOwned + JsonSchema>(
     ) -> Result<HandleResponse, Error>,
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
-) -> Result<*mut c_void, Error> {
+) -> Result<HandleResponse, Error> {
     let env: Vec<u8> = unsafe { consume_region(env_ptr)? };
     let msg: Vec<u8> = unsafe { consume_region(msg_ptr)? };
 
     let env: Env = from_slice(&env)?;
     let msg: T = from_slice(&msg)?;
     let mut deps = make_dependencies();
-    let res = handle_fn(&mut deps, env, msg)?;
-    let json = to_vec(&HandleResult::Ok(res))?;
-    Ok(release_buffer(json))
+    handle_fn(&mut deps, env, msg)
 }
 
 fn _do_query<T: DeserializeOwned + JsonSchema>(
     query_fn: &dyn Fn(&Extern<ExternalStorage, ExternalApi>, T) -> Result<QueryResponse, Error>,
     msg_ptr: *mut c_void,
-) -> Result<*mut c_void, Error> {
+) -> Result<QueryResponse, Error> {
     let msg: Vec<u8> = unsafe { consume_region(msg_ptr)? };
 
     let msg: T = from_slice(&msg)?;
     let deps = make_dependencies();
-    let res = query_fn(&deps, msg)?;
-    let json = to_vec(&QueryResult::Ok(res))?;
-    Ok(release_buffer(json))
-}
-
-fn make_init_error_c_string<T: Display>(error: T) -> *mut c_void {
-    let v = to_vec(&InitResult::Err(error.to_string())).unwrap();
-    release_buffer(v)
-}
-
-fn make_handle_error_c_string<T: Display>(error: T) -> *mut c_void {
-    let v = to_vec(&HandleResult::Err(error.to_string())).unwrap();
-    release_buffer(v)
-}
-
-fn make_query_error_c_string<T: Display>(error: T) -> *mut c_void {
-    let v = to_vec(&QueryResult::Err(error.to_string())).unwrap();
-    release_buffer(v)
+    query_fn(&deps, msg)
 }
 
 /// Makes all bridges to external dependencies (i.e. Wasm imports) that are injected by the VM