+++ packages/std/src/init_handle.rs
@@ -3,6 +3,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::api::ApiResult;
 use crate::encoding::Binary;
 use crate::types::{Coin, HumanAddr};
 
@@ -50,29 +51,7 @@ pub struct InitResponse {
     pub data: Option<Binary>,   // abci defines this as bytes
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum InitResult {
-    Ok(InitResponse),
-    Err(String),
-}
-
-impl InitResult {
-    // unwrap will panic on err, or give us the real data useful for tests
-    pub fn unwrap(self) -> InitResponse {
-        match self {
-            InitResult::Err(msg) => panic!("Unexpected error: {}", msg),
-            InitResult::Ok(res) => res,
-        }
-    }
-
-    pub fn is_err(&self) -> bool {
-        match self {
-            InitResult::Err(_) => true,
-            _ => false,
-        }
-    }
-}
+pub type InitResult = ApiResult<InitResponse>;
 
 #[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
 pub struct HandleResponse {
@@ -82,38 +61,17 @@ pub struct HandleResponse {
     pub data: Option<Binary>,   // abci defines this as bytes
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum HandleResult {
-    Ok(HandleResponse),
-    Err(String),
-}
-
-impl HandleResult {
-    // unwrap will panic on err, or give us the real data useful for tests
-    pub fn unwrap(self) -> HandleResponse {
-        match self {
-            HandleResult::Err(msg) => panic!("Unexpected error: {}", msg),
-            HandleResult::Ok(res) => res,
-        }
-    }
-
-    pub fn is_err(&self) -> bool {
-        match self {
-            HandleResult::Err(_) => true,
-            _ => false,
-        }
-    }
-}
+pub type HandleResult = ApiResult<HandleResponse>;
 
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate::api::ApiError;
     use crate::{coin, from_slice, to_vec};
 
     #[test]
     fn can_deser_error_result() {
-        let fail = InitResult::Err("foobar".to_string());
+        let fail = InitResult::Err(ApiError::Unauthorized {});
         let bin = to_vec(&fail).expect("encode contract result");
         println!("error: {}", std::str::from_utf8(&bin).unwrap());
         let back: InitResult = from_slice(&bin).expect("decode contract result");