+++ packages/std/src/query.rs
@@ -1,34 +1,13 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::api::ApiResult;
 use crate::encoding::Binary;
 use crate::types::{Coin, HumanAddr};
 
 pub type QueryResponse = Binary;
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum QueryResult {
-    Ok(QueryResponse),
-    Err(String),
-}
-
-impl QueryResult {
-    // unwrap will panic on err, or give us the real data useful for tests
-    pub fn unwrap(self) -> QueryResponse {
-        match self {
-            QueryResult::Err(msg) => panic!("Unexpected error: {}", msg),
-            QueryResult::Ok(res) => res,
-        }
-    }
-
-    pub fn is_err(&self) -> bool {
-        match self {
-            QueryResult::Err(_) => true,
-            _ => false,
-        }
-    }
-}
+pub type QueryResult = ApiResult<QueryResponse>;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]