+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - All serialized JSON types now use snake_case mappings for names. This means
   enum fields like `ChangeOwner` will map to `change_owner` in the underlying
   JSON, not `changeowner`. This is a breaking change for the clients.
+- Public interface between contract and runtime no longer uses `String` to represent
+  an error, but rather serializes `ApiError` as a rich JSON error.
 
 **cosmwasm**
 
@@ -54,6 +56,7 @@
   types).
 - Add `Querier` trait and `QueryRequest` for future query callbacks from the
   contract, along with `SystemError` type for the runtime rejecting messages.
+- `{Init,Handle,Query}Result` are now just aliases for a concrete `ApiResult` type.
 
 **cosmwasm-vm**
 
@@ -70,6 +73,8 @@
 - Provide custom `serde::{from_slice, to_vec}` implementation separate from
   `cosmwasm_std`, so we can return cosmwasm-vm specific `Result` (only used
   internally).
+- `call_{init,handle,query}` and the `cosmwasm_vm::testing` wrappers return
+  standard `Result` types now, eg. `Result<HandleResponse, ApiError>`.
 
 ## 0.7.2 (2020-03-23)